Move hook to core-data
